#29 Redux auth - Character page now working with redux
