+++ src/services/RequestService.js
@@ -1,14 +1,17 @@
-import AuthService from "./AuthService";
+// import AuthService from "./AuthService";
+import { userService } from './UserService';
 
 export default class RequestService {
   // Initializing important variables
   constructor(domain) {
     this.domain = domain || 'https://127.0.0.1:8052' // API server domain
     this.fetch = this.fetch.bind(this) // React binding stuff
-    this.Auth = new AuthService();
+    // this.Auth = new AuthService();
   }
 
   getServers = () => {
+    console.log("hello", userService.loggedIn());
+    console.log("token", userService.getToken());
     return this.fetch(`${this.domain}/realms`, {
       method: 'GET'
     }).then(res => {
@@ -35,8 +38,8 @@ export default class RequestService {
 
     // Setting Authorization header
     // Authorization: Bearer xxxxxxx.xxxxxxxx.xxxxxx
-    if (this.Auth.loggedIn()) {
-      headers['Authorization'] = 'Bearer ' + this.Auth.getToken()
+    if (userService.loggedIn()) {
+      headers['Authorization'] = 'Bearer ' + userService.getToken()
     }
 
     return fetch(url, {
