+++ src/store/configureStore.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware, combineReducers } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import alertReducer from './reducers/alert';
-import authReducer from './reducers/authentication';
+import authReducer from './reducers/user';
 
 const reducers = {
   alert: alertReducer,
