+++ src/store/reducers/user.js
@@ -5,15 +5,7 @@ const initialState = user ? { loggedIn: true, user } : {};
 
 export default function authentication(state = initialState, action) {
   switch (action.type) {
-    case userConstants.LOGIN_REQUEST:
-      console.log('login_request:action.user:', action.user);
-      return {
-        ...state,
-        loggingIn: true,
-        user: action.user
-      };
     case userConstants.LOGIN_SUCCESS:
-      console.log('login_success:action.user:', action.user);
       return {
         ...state,
         loggedIn: true,
@@ -23,12 +15,6 @@ export default function authentication(state = initialState, action) {
       return {};
     case userConstants.LOGOUT:
       return {};
-    case userConstants.REGISTER_REQUEST:
-      return {
-        ...state,
-        loggedIn: true,
-        user: action.user
-      };
     case userConstants.REGISTER_SUCCESS:
       return {
         ...state,
